chore: 完成历史人流量模块

+++ detect-service/detect-flow/src/main/java/com/explore/controller/FlowController.java
@@ -39,7 +39,7 @@ public class FlowController {
 
     private final IFlowService flowService;
 
-    private final CameraClient  cameraClient;
+    private final CameraClient cameraClient;
 
     private final IWarningService warningService;
 
@@ -51,46 +51,49 @@ public class FlowController {
 
     /**
      * 保存检测数据
+     *
      * @param flow obj
      */
     @PostMapping
-    public void save(@RequestBody Flow flow){
+    public void save(@RequestBody Flow flow) {
         flowService.save(flow);
         Camera camera = cameraClient.getById(flow.getCameraId()).getData();
-        if (flow.getFlow().compareTo(camera.getWarning())>0){
+        if (flow.getFlow().compareTo(camera.getWarning().longValue()) > 0) {
             String value = cache.getIfPresent(String.valueOf(camera.getId()));
-            if (value != null){
+            if (value != null) {
                 return;
             }
             LocalDateTime now = LocalDateTime.now();
             cache.put(String.valueOf(camera.getId()), now.toString());
-            warningService.save(Warning.builder().number(flow.getFlow()).warning(camera.getWarning()).createTime(now).build());
-            log.info(" warning => flow value : {}",flow.getFlow());
+            warningService.save(Warning.builder().number(flow.getFlow().intValue()).warning(camera.getWarning()).createTime(now).build());
+            log.info(" warning => flow value : {}", flow.getFlow());
         }
     }
 
     /**
      * 获取最新记录
+     *
      * @return ServerResponse
      */
     @GetMapping
-    public ServerResponse get(@RequestParam("cid")Integer cid){
-        Flow flow = flowService.getOne(new QueryWrapper<Flow>().eq("camera_id",cid).orderByDesc("id").last("limit 0,1"));
+    public ServerResponse get(@RequestParam("cid") Integer cid) {
+        Flow flow = flowService.getOne(new QueryWrapper<Flow>().eq("camera_id", cid).orderByDesc("id").last("limit 0,1"));
         return ServerResponse.createBySuccess(flow);
     }
 
     /**
      * 根据摄像头获取检测数据
+     *
      * @param cid 摄像头id
      * @return ServerResponse
      */
     @GetMapping("/list")
     public ServerResponse getListByCid(@RequestParam("cid") Integer cid,
-                                       @RequestParam(required = false,defaultValue = "1")Integer page,
-                                       @RequestParam(required = false,defaultValue = "10")Integer limit){
+                                       @RequestParam(required = false, defaultValue = "1") Integer page,
+                                       @RequestParam(required = false, defaultValue = "10") Integer limit) {
         return ServerResponse.createBySuccess(
                 flowService.page(
-                        new Page<>(page,limit),
-                        new QueryWrapper<Flow>().eq("camera_id",cid).orderByDesc("id")));
+                        new Page<>(page, limit),
+                        new QueryWrapper<Flow>().eq("camera_id", cid).orderByDesc("id")));
     }
 }
